+++ src/pyfx/view/components/__init__.py
@@ -4,6 +4,6 @@ Different components(or urwid widgets wrapper) will be used in *Pyfx*'s TUI.
 Each component is a single unit rendered in *Pyfx*.
 """
 from .autocomplete_popup import AutoCompletePopUp
-from .help_bar import HelpBar
+from .help import HelpBar
 from .query_bar import QueryBar
 from .json_browser import JSONBrowser
