+++ src/pyfx/view/components/help/help_bar.py
