+++ src/pyfx/view/components/help/help_popup.py
